refactor(api): git github 라우트를 전용 모듈로 추출
8개 라우트를 api/git/github.py로 이동. 핸들러 본문·경로·함수명 무변경.
라우트 테이블 characterization 테스트 통과 — HTTP 표면 동일.

공용 의존성 모듈 api/git/_shared.py 신설(레시피 R3 최초 적용).
_legacy.py 의 get_github_service 정의를 _shared.py 로 승격하고 양쪽에서
참조한다. _legacy.py 가 `from ._shared import get_github_service` 로 이름을
자기 네임스페이스에 바인딩하므로, prune-merged 핸들러를 대상으로 하는
test_api_git_prune.py 의 patch("api.git._legacy.get_github_service") 7곳은
그대로 유효하다 (R4b 는 no-op).
