+++ src/backend/api/git/__init__.py
@@ -9,13 +9,14 @@
 깨진다(Global Constraints 2 위반).
 """
 
-from . import repositories
+from . import github, repositories
 from ._legacy import (
     generate_draft_commits,
     generate_draft_commits_for_project,
     router,
 )
 
+router.include_router(github.router)
 router.include_router(repositories.router)
 
 __all__ = [
