+++ src/backend/api/git/_shared.py
@@ -1,0 +1,22 @@
+"""Git API 모듈들이 공유하는 의존성.
+
+`_legacy.py` 분할 과정에서 여러 모듈이 함께 쓰게 된 이름을 여기로 승격한다.
+순환 import 를 막기 위해 이 모듈은 형제 모듈(`._legacy` 포함)을 import 하지
+않는다 — 의존은 항상 한 방향(형제 → `_shared`)이다.
+"""
+
+from fastapi import HTTPException
+
+
+def get_github_service():
+    """Get GitHubService instance."""
+    from services.github_service import get_github_service as factory
+
+    service = factory()
+    if not service:
+        raise HTTPException(
+            status_code=503,
+            detail="GitHub service not available. Check GITHUB_TOKEN environment variable.",
+        )
+
+    return service
